Move Tester component. Add getQuery to allow for actual searching
