+++ src/helpers/getQuery.js
@@ -1,0 +1,8 @@
+export default function getQuery() {
+  try {
+    const param = new URLSearchParams(window?.location?.search)
+    return param.get('query') || undefined
+  } catch (err) {
+    return undefined
+  }
+}
